# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * F1 Intelligence API - Race prediction and analytics platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  model_ready?: boolean;
  data_cached?: boolean;
}

export interface Race {
  round: number;
  name: string;
  circuit: string;
  country: string;
  date: string;
  year: number;
  /** @nullable */
  locality?: string | null;
}

export interface Driver {
  code: string;
  name: string;
  team: string;
  number: number;
  nationality: string;
  /** @nullable */
  teamColor?: string | null;
}

export interface Constructor {
  id: string;
  name: string;
  nationality: string;
  /** @nullable */
  color?: string | null;
}

export interface PredictionRequest {
  year: number;
  round: number;
  /** @nullable */
  driver?: string | null;
  /** @nullable */
  constructor?: string | null;
}

export interface PodiumEntry {
  position: number;
  driver: string;
  team: string;
  confidence: number;
  /** @nullable */
  teamColor?: string | null;
}

export interface FeatureFactor {
  feature: string;
  importance: number;
  description: string;
}

export interface PredictionResult {
  year: number;
  round: number;
  raceName: string;
  circuit: string;
  predictedWinner: PodiumEntry;
  podium: PodiumEntry[];
  top10: PodiumEntry[];
  overallConfidence: number;
  featureFactors: FeatureFactor[];
  modelType: string;
  whyWin: string[];
  /** @nullable */
  strategyNote?: string | null;
}

export interface DriverRaceResult {
  round: number;
  raceName: string;
  /** @nullable */
  position: number | null;
  points: number;
  team: string;
  /** @nullable */
  teamColor?: string | null;
  fastestLap?: boolean;
}

export interface ConstructorStandingsPoint {
  round: number;
  raceName: string;
  constructor: string;
  points: number;
  /** @nullable */
  color: string | null;
}

export interface DriverStanding {
  position: number;
  driver: string;
  team: string;
  points: number;
  wins: number;
  nationality?: string;
  /** @nullable */
  teamColor?: string | null;
}

export interface WinStat {
  driver: string;
  wins: number;
  races: number;
  winPct: number;
  team: string;
  /** @nullable */
  teamColor?: string | null;
}

export interface PoleStat {
  driver: string;
  poles: number;
  team: string;
  /** @nullable */
  teamColor?: string | null;
}

export interface CircuitStat {
  circuit: string;
  country: string;
  avgPosition: number;
  bestPosition: number;
  races: number;
  wins?: number;
}

export interface AvgFinishStat {
  driver: string;
  avgPosition: number;
  team: string;
  races: number;
  /** @nullable */
  teamColor?: string | null;
}

export interface DatasetRow {
  year: number;
  round: number;
  raceName: string;
  circuit: string;
  country: string;
  driver: string;
  driverCode?: string;
  team: string;
  /** @nullable */
  position: number | null;
  points: number;
  /** @nullable */
  gridPosition: number | null;
  fastestLap?: boolean;
  /** @nullable */
  lapsCompleted?: number | null;
  status?: string;
}

export interface DatasetResponse {
  total: number;
  rows: DatasetRow[];
}

export type ListRacesParams = {
/**
 * Season year (defaults to 2024)
 */
year?: number;
};

export type ListDriversParams = {
year?: number;
};

export type ListConstructorsParams = {
year?: number;
};

export type GetDriverPerformanceParams = {
/**
 * Driver code e.g. VER
 */
driver: string;
year?: number;
};

export type GetConstructorStandingsParams = {
year?: number;
};

export type GetDriverStandingsParams = {
year?: number;
};

export type GetWinPercentageParams = {
year?: number;
};

export type GetPolePositionsParams = {
year?: number;
};

export type GetCircuitPerformanceParams = {
driver: string;
year?: number;
};

export type GetAvgFinishParams = {
year?: number;
};

export type GetDatasetParams = {
/**
 * Search term for driver, team, or circuit
 */
search?: string;
year?: number;
driver?: string;
constructor?: string;
limit?: number;
offset?: number;
};