# lib/api-zod/src/generated/types/avgFinishStat.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * F1 Intelligence API - Race prediction and analytics platform
 * OpenAPI spec version: 0.1.0
 */

export interface AvgFinishStat {
  driver: string;
  avgPosition: number;
  team: string;
  races: number;
  /** @nullable */
  teamColor?: string | null;
}