# lib/api-zod/src/generated/types/circuitStat.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * F1 Intelligence API - Race prediction and analytics platform
 * OpenAPI spec version: 0.1.0
 */

export interface CircuitStat {
  circuit: string;
  country: string;
  avgPosition: number;
  bestPosition: number;
  races: number;
  wins?: number;
}