# lib/api-zod/src/generated/types/constructorStandingsPoint.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * F1 Intelligence API - Race prediction and analytics platform
 * OpenAPI spec version: 0.1.0
 */

export interface ConstructorStandingsPoint {
  round: number;
  raceName: string;
  constructor: string;
  points: number;
  /** @nullable */
  color: string | null;
}