# lib/api-zod/src/generated/types/datasetResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * F1 Intelligence API - Race prediction and analytics platform
 * OpenAPI spec version: 0.1.0
 */
import type { DatasetRow } from './datasetRow';

export interface DatasetResponse {
  total: number;
  rows: DatasetRow[];
}