# lib/api-zod/src/generated/types/datasetRow.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * F1 Intelligence API - Race prediction and analytics platform
 * OpenAPI spec version: 0.1.0
 */

export interface DatasetRow {
  year: number;
  round: number;
  raceName: string;
  circuit: string;
  country: string;
  driver: string;
  driverCode?: string;
  team: string;
  /** @nullable */
  position: number | null;
  points: number;
  /** @nullable */
  gridPosition: number | null;
  fastestLap?: boolean;
  /** @nullable */
  lapsCompleted?: number | null;
  status?: string;
}