# lib/api-zod/src/generated/types/driverRaceResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * F1 Intelligence API - Race prediction and analytics platform
 * OpenAPI spec version: 0.1.0
 */

export interface DriverRaceResult {
  round: number;
  raceName: string;
  /** @nullable */
  position: number | null;
  points: number;
  team: string;
  /** @nullable */
  teamColor?: string | null;
  fastestLap?: boolean;
}