# lib/api-zod/src/generated/types/podiumEntry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * F1 Intelligence API - Race prediction and analytics platform
 * OpenAPI spec version: 0.1.0
 */

export interface PodiumEntry {
  position: number;
  driver: string;
  team: string;
  confidence: number;
  /** @nullable */
  teamColor?: string | null;
}