# lib/api-zod/src/generated/types/predictionRequest.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * F1 Intelligence API - Race prediction and analytics platform
 * OpenAPI spec version: 0.1.0
 */

export interface PredictionRequest {
  year: number;
  round: number;
  /** @nullable */
  driver?: string | null;
  /** @nullable */
  constructor?: string | null;
}